requirements.txt
# Core dependencies
pg8000>=1.31.0
boto3>=1.34.0
python-dotenv>=1.0.0

# For PDF processing
pypdf>=4.0.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0